refactor: move offchain logic to separate file
